requirements.txt
diffusers==0.31.0
transformers==4.44.0
accelerate>=0.33.0
safetensors>=0.4.0
huggingface-hub>=0.24.0
einops==0.8.0
numpy==1.26.4
omegaconf==2.3.0
pydantic==2.8.2
Pillow==10.4.0
opencv-python==4.11.0.86
kornia==0.7.3
scikit-learn==1.5.2
pyamg==5.2.1
matplotlib==3.9.1.post1
tqdm==4.66.5
openai==1.79.0
python-dotenv==1.2.1
